# docs/hai-contract/openapi.yaml
openapi: 3.0.0
info:
  title: AGORA HAI REST API
  version: 2.0.0
  description: |
    REST API for the AGORA Human Agent Interface (HAI).
    This API handles session management, conversation history retrieval, and other non-real-time operations
    that complement the AG-UI WebSocket protocol.

  contact:
    name: AGORA Development Team
    email: lex@gradient-ds.com
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /sessions:
    get:
      summary: List sessions for a user
      description: |
        Retrieves all sessions for a specific user, ordered by last activity (most recent first).
        Used by the frontend to display the conversation history sidebar.
      operationId: listSessions
      tags:
        - Sessions
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            example: "koen"
          required: true
          description: Inspector persona ID (e.g., "koen", "fatima", "jan")
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Maximum number of sessions to return
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Pagination offset
      responses:
        '200':
          description: Session list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{session_id}:
    put:
      summary: Update a session
      description: |
        Updates session metadata such as the title. Useful for allowing users to rename conversations.
      operationId: updateSession
      tags:
        - Sessions
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
          required: true
          description: Unique session identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionRequest'
      responses:
        '200':
          description: Session updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  session:
                    $ref: '#/components/schemas/SessionMetadata'
        '400':
          description: No update fields provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a session
      description: |
        Deletes a session and all its associated data, including conversation history and metadata.
      operationId: deleteSession
      tags:
        - Sessions
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
          required: true
          description: Unique session identifier
      responses:
        '200':
          description: Session deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Session deleted"
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{session_id}/metadata:
    get:
      summary: Get session metadata
      description: |
        Retrieves metadata for a specific session without the full conversation history.
        Useful for displaying session info in the sidebar without loading all messages.
      operationId: getSessionMetadata
      tags:
        - Sessions
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
          required: true
          description: Unique session identifier
      responses:
        '200':
          description: Session metadata retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  session:
                    $ref: '#/components/schemas/SessionMetadata'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{session_id}/history:
    get:
      summary: Get conversation history
      description: |
        Retrieves the full conversation history for a specific session.
        This should be called by the client when switching sessions or on page load
        before establishing the WebSocket connection.
      operationId: getSessionHistory
      tags:
        - Sessions
      parameters:
        - in: path
          name: session_id
          schema:
            type: string
            example: "123e4567-e89b-12d3-a456-426614174000"
          required: true
          description: Unique session identifier (same as threadId)
        - in: query
          name: include_tools
          schema:
            type: boolean
            default: false
          description: |
            If true, includes tool calls (role: "tool_call") and tool results (role: "tool") in the history.
            Required for restoring the debug panel state.
      responses:
        '200':
          description: Conversation history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------------
  # USER MANAGEMENT
  # ---------------------------------------------------------------------------
  /users/me:
    get:
      summary: Get current user profile
      description: Retrieves profile information for the specified user.
      operationId: getCurrentUser
      tags:
        - Users
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
          required: true
          description: Current user ID
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me/preferences:
    get:
      summary: Get user preferences
      description: Retrieves the current user's preferences.
      operationId: getUserPreferences
      tags:
        - Users
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
          required: true
          description: Current user ID
      responses:
        '200':
          description: Preferences retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  preferences:
                    $ref: '#/components/schemas/UserPreferences'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update user preferences
      description: Updates settings like spoken response mode.
      operationId: updateUserPreferences
      tags:
        - Users
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
          required: true
          description: Current user ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
      responses:
        '200':
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  preferences:
                    $ref: '#/components/schemas/UserPreferences'
        '400':
          description: Invalid input (e.g., invalid spoken_response_mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users:
    post:
      summary: Create a new user
      description: |
        Creates a new user account. Email must be unique across the system.
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid input (e.g., missing required fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List all users
      description: |
        Retrieves a paginated list of all users, ordered by creation date (most recent first).
        Requires admin privileges.
      operationId: listUsers
      tags:
        - Users
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Maximum number of users to return
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Pagination offset
      responses:
        '200':
          description: User list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    UpdateSessionRequest:
      type: object
      description: Request body for updating session metadata
      properties:
        title:
          type: string
          maxLength: 200
          description: New session title
    SessionMetadata:
      type: object
      description: Metadata for a conversation session
      required: [sessionId, userId, title, messageCount, createdAt, lastActivity]
      properties:
        sessionId:
          type: string
          description: Unique session identifier (same as threadId)
          example: "123e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          description: Inspector persona ID
          example: "koen"
        title:
          type: string
          description: Auto-generated title from first message
          example: "Inspectie bij Restaurant Bella Rosa"
        firstMessagePreview:
          type: string
          description: Preview of the first message (truncated)
          example: "Start inspectie bij Restaurant..."
        messageCount:
          type: integer
          description: Total number of messages in the session
          example: 12
        createdAt:
          type: string
          format: date-time
          description: Session creation timestamp
          example: "2025-12-01T10:30:00Z"
        lastActivity:
          type: string
          format: date-time
          description: Last message timestamp
          example: "2025-12-01T11:45:00Z"
    SessionListResponse:
      type: object
      description: Response for listing sessions
      required: [success, sessions, totalCount]
      properties:
        success:
          type: boolean
          example: true
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionMetadata'
        totalCount:
          type: integer
          description: Total number of sessions (for pagination)
          example: 42

    # -------------------------------------------------------------------------
    # HISTORY SCHEMAS
    # -------------------------------------------------------------------------
    HistoryResponse:
      type: object
      description: Response for conversation history
      required: [success, threadId, history, messageCount]
      properties:
        success:
          type: boolean
          example: true
        threadId:
          type: string
          description: Session/thread identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        history:
          type: array
          items:
            $ref: '#/components/schemas/HistoryMessage'
        messageCount:
          type: integer
          description: Number of messages in history
          example: 15
...
    UserPreferences:
      type: object
      description: User preferences stored in the system
      properties:
        theme:
          type: string
          enum: [light, dark, system]
          default: system
          description: UI theme preference
        notifications_enabled:
          type: boolean
          default: true
          description: Enable in-app notifications
        default_agent_id:
          type: string
          default: "general-agent"
          example: "general-agent"
          description: Default agent to use for new conversations
        language:
          type: string
          default: "nl-NL"
          example: "nl-NL"
          description: UI language preference
        spoken_text_type:
          type: string
          enum: [dictate, summarize]
          default: summarize
          description: |
            Type of spoken text for TTS:
            - dictate: Full text-to-speech of responses
            - summarize: Summarized spoken responses
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error code or short type
          example: "not_found"
        message:
          type: string
          description: Human readable error message
          example: "The requested resource could not be found."